[Decomp][Player] Cohesion split: extract instance reset/difficulty senders to PlayerInstance.cpp
