+++ src/game/Object/Player.cpp
@@ -5364,114 +5364,6 @@ void Player::SendExplorationExperience(uint32 Area, uint32 Experience)
     GetSession()->SendPacket(&data);
 }
 
-void Player::SendDungeonDifficulty(bool IsInGroup)
-{
-    uint8 val = 0x00000001;
-    WorldPacket data(MSG_SET_DUNGEON_DIFFICULTY, 12);
-    data << uint32(GetDungeonDifficulty());
-    data << uint32(val);
-    data << uint32(IsInGroup);
-    GetSession()->SendPacket(&data);
-}
-
-void Player::SendRaidDifficulty(bool IsInGroup)
-{
-    uint8 val = 0x00000001;
-    WorldPacket data(MSG_SET_RAID_DIFFICULTY, 12);
-    data << uint32(GetRaidDifficulty());
-    data << uint32(val);
-    data << uint32(IsInGroup);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends a reset-failed notification for an instance map.
- *
- * @param mapid The map identifier that failed to reset.
- */
-void Player::SendResetFailedNotify(uint32 mapid)
-{
-    WorldPacket data(SMSG_RESET_FAILED_NOTIFY, 4);
-    data << uint32(mapid);
-    GetSession()->SendPacket(&data);
-}
-
-/// Reset all solo instances and optionally send a message on success for each
-void Player::ResetInstances(InstanceResetMethod method, bool isRaid)
-{
-    // method can be INSTANCE_RESET_ALL, INSTANCE_RESET_CHANGE_DIFFICULTY, INSTANCE_RESET_GROUP_JOIN
-
-    // we assume that when the difficulty changes, all instances that can be reset will be
-    Difficulty diff = GetDifficulty(isRaid);
-
-    for (BoundInstancesMap::iterator itr = m_boundInstances[diff].begin(); itr != m_boundInstances[diff].end();)
-    {
-        DungeonPersistentState* state = itr->second.state;
-        const MapEntry* entry = sMapStore.LookupEntry(itr->first);
-        if (!entry || entry->IsRaid() != isRaid || !state->CanReset())
-        {
-            ++itr;
-            continue;
-        }
-
-        if (method == INSTANCE_RESET_ALL)
-        {
-            // the "reset all instances" method can only reset normal maps
-            if (entry->map_type == MAP_RAID || diff == DUNGEON_DIFFICULTY_HEROIC)
-            {
-                ++itr;
-                continue;
-            }
-        }
-
-        // if the map is loaded, reset it
-        if (Map* map = sMapMgr.FindMap(state->GetMapId(), state->GetInstanceId()))
-            if (map->IsDungeon())
-            {
-                ((DungeonMap*)map)->Reset(method);
-            }
-
-        // since this is a solo instance there should not be any players inside
-        if (method == INSTANCE_RESET_ALL || method == INSTANCE_RESET_CHANGE_DIFFICULTY)
-        {
-            SendResetInstanceSuccess(state->GetMapId());
-        }
-
-        state->DeleteFromDB();
-        m_boundInstances[diff].erase(itr++);
-
-        // the following should remove the instance save from the manager and delete it as well
-        state->RemovePlayer(this);
-    }
-}
-
-/**
- * @brief Sends a successful instance reset notification to the client.
- *
- * @param MapId The reset map identifier.
- */
-void Player::SendResetInstanceSuccess(uint32 MapId)
-{
-    WorldPacket data(SMSG_INSTANCE_RESET, 4);
-    data << uint32(MapId);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends an instance reset failure message to the client.
- *
- * @param reason The reset failure reason code.
- * @param MapId The map identifier that failed to reset.
- */
-void Player::SendResetInstanceFailed(uint32 reason, uint32 MapId)
-{
-    // TODO: find what other fail reasons there are besides players in the instance
-    WorldPacket data(SMSG_INSTANCE_RESET_FAILED, 4);
-    data << uint32(reason);
-    data << uint32(MapId);
-    GetSession()->SendPacket(&data);
-}
-
 /*********************************************************/
 /***              Update timers                        ***/
 /*********************************************************/
@@ -6705,46 +6597,6 @@ void Player::SendTransferAbortedByLockStatus(MapEntry const* mapEntry, AreaLockS
             sLog.outError("SendTransfertAbortedByLockstatus: unhandled LockAreaStatus %u, when %s attempts to enter in map %u", lockStatus, GetGuidStr().c_str(), mapEntry->MapID);
             break;
     }
-}
-
-/**
- * @brief Sends an instance reset warning message to the client.
- *
- * @param mapid The map identifier being reset.
- * @param time The remaining time until reset in seconds.
- */
-void Player::SendInstanceResetWarning(uint32 mapid, Difficulty difficulty, uint32 time)
-{
-    // type of warning, based on the time remaining until reset
-    uint32 type;
-    if (time > 3600)
-    {
-        type = RAID_INSTANCE_WELCOME;
-    }
-    else if (time > 900 && time <= 3600)
-    {
-        type = RAID_INSTANCE_WARNING_HOURS;
-    }
-    else if (time > 300 && time <= 900)
-    {
-        type = RAID_INSTANCE_WARNING_MIN;
-    }
-    else
-    {
-        type = RAID_INSTANCE_WARNING_MIN_SOON;
-    }
-
-    WorldPacket data(SMSG_RAID_INSTANCE_MESSAGE, 4 + 4 + 4 + 4);
-    data << uint32(type);
-    data << uint32(mapid);
-    data << uint32(difficulty);                             // difficulty
-    data << uint32(time);
-    if (type == RAID_INSTANCE_WELCOME)
-    {
-        data << uint8(0);                                   // is your (1)
-        data << uint8(0);                                   // is extended (1), ignored if prev field is 0
-    }
-    GetSession()->SendPacket(&data);
 }
 
 /**
