+++ src/game/Object/PlayerInstance.cpp
@@ -1,0 +1,229 @@
+/**
+ * MaNGOS is a full featured server for World of Warcraft, supporting
+ * the following clients: 1.12.x, 2.4.3, 3.3.5a, 4.3.4a and 5.4.8
+ *
+ * Copyright (C) 2005-2025 MaNGOS <https://www.getmangos.eu>
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ *
+ * World of Warcraft, and all World of Warcraft or Warcraft art, images,
+ * and lore are copyrighted by Blizzard Entertainment, Inc.
+ */
+
+#include "Player.h"
+#include "Language.h"
+#include "Database/DatabaseEnv.h"
+#include "Log.h"
+#include "Opcodes.h"
+#include "SpellMgr.h"
+#include "World.h"
+#include "WorldPacket.h"
+#include "WorldSession.h"
+#include "UpdateMask.h"
+#include "SkillDiscovery.h"
+#include "QuestDef.h"
+#include "GossipDef.h"
+#include "UpdateData.h"
+#include "Channel.h"
+#include "ChannelMgr.h"
+#include "MapManager.h"
+#include "MapPersistentStateMgr.h"
+#include "InstanceData.h"
+#include "GridNotifiers.h"
+#include "GridNotifiersImpl.h"
+#include "CellImpl.h"
+#include "ObjectMgr.h"
+#include "ObjectAccessor.h"
+#include "CreatureAI.h"
+#include "Formulas.h"
+#include "Group.h"
+#include "Guild.h"
+#include "GuildMgr.h"
+#include "Pet.h"
+#include "Util.h"
+#include "Transports.h"
+#include "Weather.h"
+#include "BattleGround/BattleGround.h"
+#include "BattleGround/BattleGroundMgr.h"
+#include "BattleGround/BattleGroundAV.h"
+#include "OutdoorPvP/OutdoorPvP.h"
+#include "ArenaTeam.h"
+#include "Chat.h"
+#include "revision_data.h"
+#include "Database/DatabaseImpl.h"
+#include "Spell.h"
+#include "ScriptMgr.h"
+#include "SocialMgr.h"
+#include "AchievementMgr.h"
+#include "Mail.h"
+#include "SpellAuras.h"
+#include "DBCStores.h"
+#include "DB2Stores.h"
+#include "SQLStorages.h"
+#include "Vehicle.h"
+#include "Calendar.h"
+#include "DisableMgr.h"
+#ifdef ENABLE_ELUNA
+#include "LuaEngine.h"
+#endif /* ENABLE_ELUNA */
+
+#include <cmath>
+void Player::SendDungeonDifficulty(bool IsInGroup)
+{
+    uint8 val = 0x00000001;
+    WorldPacket data(MSG_SET_DUNGEON_DIFFICULTY, 12);
+    data << uint32(GetDungeonDifficulty());
+    data << uint32(val);
+    data << uint32(IsInGroup);
+    GetSession()->SendPacket(&data);
+}
+
+void Player::SendRaidDifficulty(bool IsInGroup)
+{
+    uint8 val = 0x00000001;
+    WorldPacket data(MSG_SET_RAID_DIFFICULTY, 12);
+    data << uint32(GetRaidDifficulty());
+    data << uint32(val);
+    data << uint32(IsInGroup);
+    GetSession()->SendPacket(&data);
+}
+
+/**
+ * @brief Sends a reset-failed notification for an instance map.
+ *
+ * @param mapid The map identifier that failed to reset.
+ */
+void Player::SendResetFailedNotify(uint32 mapid)
+{
+    WorldPacket data(SMSG_RESET_FAILED_NOTIFY, 4);
+    data << uint32(mapid);
+    GetSession()->SendPacket(&data);
+}
+
+/// Reset all solo instances and optionally send a message on success for each
+void Player::ResetInstances(InstanceResetMethod method, bool isRaid)
+{
+    // method can be INSTANCE_RESET_ALL, INSTANCE_RESET_CHANGE_DIFFICULTY, INSTANCE_RESET_GROUP_JOIN
+
+    // we assume that when the difficulty changes, all instances that can be reset will be
+    Difficulty diff = GetDifficulty(isRaid);
+
+    for (BoundInstancesMap::iterator itr = m_boundInstances[diff].begin(); itr != m_boundInstances[diff].end();)
+    {
+        DungeonPersistentState* state = itr->second.state;
+        const MapEntry* entry = sMapStore.LookupEntry(itr->first);
+        if (!entry || entry->IsRaid() != isRaid || !state->CanReset())
+        {
+            ++itr;
+            continue;
+        }
+
+        if (method == INSTANCE_RESET_ALL)
+        {
+            // the "reset all instances" method can only reset normal maps
+            if (entry->map_type == MAP_RAID || diff == DUNGEON_DIFFICULTY_HEROIC)
+            {
+                ++itr;
+                continue;
+            }
+        }
+
+        // if the map is loaded, reset it
+        if (Map* map = sMapMgr.FindMap(state->GetMapId(), state->GetInstanceId()))
+            if (map->IsDungeon())
+            {
+                ((DungeonMap*)map)->Reset(method);
+            }
+
+        // since this is a solo instance there should not be any players inside
+        if (method == INSTANCE_RESET_ALL || method == INSTANCE_RESET_CHANGE_DIFFICULTY)
+        {
+            SendResetInstanceSuccess(state->GetMapId());
+        }
+
+        state->DeleteFromDB();
+        m_boundInstances[diff].erase(itr++);
+
+        // the following should remove the instance save from the manager and delete it as well
+        state->RemovePlayer(this);
+    }
+}
+
+/**
+ * @brief Sends a successful instance reset notification to the client.
+ *
+ * @param MapId The reset map identifier.
+ */
+void Player::SendResetInstanceSuccess(uint32 MapId)
+{
+    WorldPacket data(SMSG_INSTANCE_RESET, 4);
+    data << uint32(MapId);
+    GetSession()->SendPacket(&data);
+}
+
+/**
+ * @brief Sends an instance reset failure message to the client.
+ *
+ * @param reason The reset failure reason code.
+ * @param MapId The map identifier that failed to reset.
+ */
+void Player::SendResetInstanceFailed(uint32 reason, uint32 MapId)
+{
+    // TODO: find what other fail reasons there are besides players in the instance
+    WorldPacket data(SMSG_INSTANCE_RESET_FAILED, 4);
+    data << uint32(reason);
+    data << uint32(MapId);
+    GetSession()->SendPacket(&data);
+}
+
+/**
+ * @brief Sends an instance reset warning message to the client.
+ *
+ * @param mapid The map identifier being reset.
+ * @param time The remaining time until reset in seconds.
+ */
+void Player::SendInstanceResetWarning(uint32 mapid, Difficulty difficulty, uint32 time)
+{
+    // type of warning, based on the time remaining until reset
+    uint32 type;
+    if (time > 3600)
+    {
+        type = RAID_INSTANCE_WELCOME;
+    }
+    else if (time > 900 && time <= 3600)
+    {
+        type = RAID_INSTANCE_WARNING_HOURS;
+    }
+    else if (time > 300 && time <= 900)
+    {
+        type = RAID_INSTANCE_WARNING_MIN;
+    }
+    else
+    {
+        type = RAID_INSTANCE_WARNING_MIN_SOON;
+    }
+
+    WorldPacket data(SMSG_RAID_INSTANCE_MESSAGE, 4 + 4 + 4 + 4);
+    data << uint32(type);
+    data << uint32(mapid);
+    data << uint32(difficulty);                             // difficulty
+    data << uint32(time);
+    if (type == RAID_INSTANCE_WELCOME)
+    {
+        data << uint8(0);                                   // is your (1)
+        data << uint8(0);                                   // is extended (1), ignored if prev field is 0
+    }
+    GetSession()->SendPacket(&data);
+}
